requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.8.0
dnspython==2.4.2
duckduckgo-search==3.8.5
exceptiongroup==1.1.3
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.0
gitdb==4.0.10
GitPython==3.1.36
greenlet==2.0.2
grpcio==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.345
langchain-core==0.0.9
langchain-experimental==0.0.43
langsmith==0.0.69
llama-index==0.8.69.post1
loguru==0.7.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
momento==1.14.1
momento-wire-types==0.96.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
openai==1.3.7
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
pinecone-client==2.2.4
protobuf==4.24.3
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pypdf==3.17.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
reportlab==4.0.7
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
safetensors==0.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.20
streamlit==1.29.0
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.2
zipp==3.16.2
